+++ src/Commands/Databases/Create.php
@@ -23,7 +23,7 @@ class Create extends BaseCommand
      */
     public function configure()
     {
-        $this->setName('create:database')
+        $this->setName('make:database')
             ->addArgument('server', InputArgument::REQUIRED, 'The id of the server to create the database on.')
             ->addOption('user', null, InputOption::VALUE_REQUIRED, 'The username of an (optional) database user to create.', null)
             ->addOption('password', null, InputOption::VALUE_REQUIRED, 'The password of the database user. Required with "user".', null)