+++ src/Commands/Servers/Create.php
@@ -30,7 +30,7 @@ class Create extends BaseCommand
      */
     public function configure()
     {
-        $this->setName('create:server')
+        $this->setName('make:server')
             ->addOption('provider', null, InputOption::VALUE_REQUIRED, 'The provider to provision the server on. Can be either "ocean2", "linode", "aws", or "custom".', 'ocean2')
             ->addOption('credential', 'C', InputOption::VALUE_REQUIRED, 'Credentials to use. Not required when using the "custom" provider.')
             ->addOption('region', null, InputOption::VALUE_REQUIRED, 'The region to provision the server in. Not required when using the "custom" provider.')